+++ drep/d_bonus.py
@@ -98,7 +98,7 @@ def check_dependencies(wd, **kwargs):
     For all possible dependencies, see if you can find them
     '''
     for dep in ['mash', 'nucmer', 'checkm', 'ANIcalculator', 'prodigal', 'centrifuge',
-                'nsimscan']:
+                'nsimscan', 'fastANI']:
         loc, works = find_program(dep)
         works_message = {True:'all good', False:'!!! ERROR !!!'}[works]
         logging.info('{0:.<40} {1:15} (location = {2})'.format(dep, works_message, loc))